backend/requirements.txt
Django==3.0.7
gunicorn
Pillow
psycopg2
djangorestframework
python-decouple